Add initial leaky bucket implementation
