+++ README.md
@@ -30,7 +30,7 @@ You must [sign up for/create a Discogs app](https://www.discogs.com/settings/dev
 * **SHIPS_FROM** (req) - Country where Marketplace listings should ship from (e.g. Australia or United States - case insensitive)
 * **DESTINATION_EMAIL** (req) - Destination email address to send digest
 * **SENDER_EMAIL** (req) - Email address to send digest from via SendGrid (must be configured via SendGrid)
-* **LOG_WANTLIST** (opt) - If true, user's wantlist will be logged to console
+* **DEBUG** (opt) - If true, enables debug logging to console
 
 **All required environment variables above must be set before `yarn run dev` command. These can also be set via a .env file.**
