+++ package.json
@@ -12,28 +12,31 @@
   "author": "Jordan Hornblow <jordan@jch254.com>",
   "license": "MIT",
   "dependencies": {
-    "@sendgrid/mail": "^7.7.0",
+    "@sendgrid/mail": "^8.1.1",
     "disconnect": "git+https://github.com/jch254/disconnect.git",
+    "fix-esm": "^1.0.1",
+    "leaky-bucket": "^4.1.4",
     "lodash": "^4.17.21",
+    "node": "^18.20.0",
     "rss-parser": "^3.13.0"
   },
   "devDependencies": {
-    "@types/aws-lambda": "^8.10.119",
-    "@types/node": "^20.4.4",
-    "@types/webpack": "^5.28.1",
-    "@types/webpack-node-externals": "^3.0.0",
+    "@types/aws-lambda": "^8.10.136",
+    "@types/node": "^20.12.2",
+    "@types/webpack": "^5.28.5",
+    "@types/webpack-node-externals": "^3.0.4",
     "cross-env": "^7.0.3",
-    "serverless": "^3.33.0",
-    "serverless-offline": "^12.0.4",
-    "serverless-webpack": "^5.11.0",
-    "source-map-loader": "^4.0.1",
-    "ts-loader": "^9.4.4",
+    "serverless": "^3.38.0",
+    "serverless-offline": "^13.3.3",
+    "serverless-webpack": "^5.13.0",
+    "source-map-loader": "^5.0.0",
+    "ts-loader": "^9.5.1",
     "tslint": "^5.20.1",
     "tslint-config-airbnb": "^5.11.2",
     "tslint-loader": "^3.5.4",
     "tsutils": "^3.21.0",
-    "typescript": "^5.1.6",
-    "webpack": "^5.88.2",
+    "typescript": "^5.4.3",
+    "webpack": "^5.91.0",
     "webpack-node-externals": "^3.0.0"
   }
 }
