+++ src/globals.d.ts
@@ -1,1 +1,2 @@
 declare module 'serverless-webpack';
+declare module 'leaky-bucket';
